Import patches-unapplied version 2.6.0~git20160810.7f5f2bd-0ubuntu1 to ubuntu/yakkety-proposed
Imported using git-ubuntu import.

Changelog parent: 63cc20212a51f701180797f637aab1fb231d7f79

New changelog entries:
  * New upstream snapshot for 2.6.0 release goal.
  * d/control: Bumped Standards-Version to 3.9.8, no changes.
  * d/openvswitch-common.install: Drop ovs-benchmark tool, removed upstream
    due to lack of use and maintenance.
  * d/control,ovn-controller-vtep.*: Add ovn-controller-vtep package.
  * d/openvswitch-switch.install: Add ovs-tcpdump binary and manpage.
  * Add support for Python 3:
    - d/control: Drop surplus Python 2 BD's, add BD's for Python 3.
    - d/control: Add missing dependency on python-six.
    - d/rules,control: Add python3-openvswitch package.
  * d/rules: Skip racey/failing tests until resolved upstream.

+++ python/ovstest/args.py
@@ -24,6 +24,7 @@ import sys
 CONTROL_PORT = 15531
 DATA_PORT = 15532
 
+
 def ip_address(string):
     """Verifies if string is a valid IP address"""
     try:
@@ -160,7 +161,7 @@ def bandwidth(string):
         raise argparse.ArgumentTypeError("Not a valid target bandwidth")
     bwidth = string.replace("M", "000000")
     bwidth = bwidth.replace("K", "000")
-    return long(bwidth) / 8  # Convert from bits to bytes
+    return int(bwidth) / 8  # Convert from bits to bytes
 
 
 def tunnel_types(string):
@@ -245,6 +246,7 @@ def ovs_initialize_args():
                 'ovs-test server in the client mode by using 127.0.0.1 as '
                 'OuterIP.')
     return parser.parse_args()
+
 
 def l3_initialize_args():
     """
